Add chart

+++ SalaryCounter/Views/EmployerPage.xaml.cs
@@ -1,3 +1,5 @@
+using Microsoft.Maui;
+
 namespace SalaryCounter.Views;
 
 public partial class EmployerPage : ContentPage
@@ -18,14 +20,19 @@ public partial class EmployerPage : ContentPage
     }
     private void Clicked(object sender, EventArgs e)
     {
+        if (pickerCommand.SelectedItem == null)
+            return;
         if (pickerCommand.SelectedItem.Equals("Добавить"))
         {
-            var employer = new Employer()
+            if (!FirstNameEntry.Text.Equals("") && !LastNameEntry.Text.Equals(""))
+            { 
+                var employer = new Employer()
             {
                 FirstName = FirstNameEntry.Text,
                 LastName = LastNameEntry.Text
             };
             _db.CreateEmployer(employer);
+        } else AppShell.Current.DisplayAlert("Добавление", "Есть пустые поля", "OK");
             GetEmployers();
         }
         if (pickerCommand.SelectedItem.Equals("Изменить"))
@@ -36,10 +43,13 @@ public partial class EmployerPage : ContentPage
             var employer = collectionView.SelectedItem as Employer;
             if (employer is null)
                 return;
-
-            employer.FirstName = FirstNameEntry.Text;
-            employer.LastName = LastNameEntry.Text;
-            _db.UpdateEmployer(employer);
+            if (!FirstNameEntry.Text.Equals("") && !LastNameEntry.Text.Equals(""))
+            {
+                employer.FirstName = FirstNameEntry.Text;
+                employer.LastName = LastNameEntry.Text;
+                _db.UpdateEmployer(employer);
+            }
+            else AppShell.Current.DisplayAlert("Изменение", "Есть пустые поля", "OK");
             GetEmployers();
         }
         if (pickerCommand.SelectedItem.Equals("Удалить"))
@@ -50,13 +60,31 @@ public partial class EmployerPage : ContentPage
             var employer = collectionView.SelectedItem as Employer;
             if (employer is null)
                 return;
-
-            _db.DeleteEmployer(employer);
+            if(_db.FindEmployers(employer.Id).Count()<1)
+                _db.DeleteEmployer(employer);
+            else AppShell.Current.DisplayAlert("Удаление", "Сначала удалите задачи, в которых участвует сотрудник", "OK");
             GetEmployers();
         }
     }
+    private void Search(object sender, EventArgs e)
+    {
+        collectionView.ItemsSource = _db.GetEmployers();
+        var source = new List<Employer>();
+        foreach (var item in collectionView.ItemsSource)
+        {
+            if (item is Employer collectionItem)
+            {
+                if (collectionItem.LastName.Contains(SearchEntry.Text))
+                    source.Add(collectionItem);
+            }
+        }
+        collectionView.ItemsSource = source;
+    }
     private void GetEmployers()
     {
+        FirstNameEntry.Text = string.Empty;
+        LastNameEntry.Text = string.Empty;
+        SearchEntry.Text = string.Empty;
         collectionView.ItemsSource = _db.GetEmployers();
     }
 }
